refactoring as per discussion
